OSGi Module -Generic Activators for Drivers and Transports -Add Bundle-Activator in Driver's/Transport's POM -Adapted EIP Feature generation
