+++ plc4j/pom.xml
@@ -36,13 +36,14 @@
   <modules>
     <module>api</module>
     <module>drivers</module>
-    <module>examples</module>
+    <!--module>examples</module-->
     <module>integrations</module>
-    <!--module>karaf-features</module-->
+    <module>karaf-features</module>
     <module>spi</module>
     <module>tools</module>
     <module>transports</module>
     <module>utils</module>
+    <module>osgi</module>
   </modules>
 
   <build>
