+++ plc4j/spi/pom.xml
@@ -84,12 +84,7 @@
       <groupId>commons-codec</groupId>
       <artifactId>commons-codec</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.osgi</groupId>
-      <artifactId>osgi.core</artifactId>
-      <version>6.0.0</version>
-      <scope>compile</scope>
-    </dependency>
+
   </dependencies>
 
   <build>
@@ -100,7 +95,6 @@
         <extensions>true</extensions>
         <configuration>
           <instructions>
-            <Bundle-Activator>org.apache.plc4x.java.osgi.SpiActivator</Bundle-Activator>
             <Export-package>io.netty.bootstrap,*</Export-package>
           </instructions>
         </configuration>
